dom/nodes/Node-appendChild-script-and-custom-from-fragment.html

<!doctype html>
<meta charset=utf-8>
<title>Node.appendChild: inserting script and custom element from a DocumentFragment</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<script>
let customConstructed = false;
let isCustomConstructed = false;
class CustomElement extends HTMLElement {
    constructor() {
        super();
        customConstructed = true;
    }
}
test(() => {
  const script = document.createElement("script");
  script.textContent = `
    customElements.define("custom-element", CustomElement);
    isCustomConstructed = customConstructed;
`;
  const custom = document.createElement("custom-element");
  const df = document.createDocumentFragment();
  df.appendChild(script);
  df.appendChild(custom);
  assert_false(customConstructed);
  assert_false(isCustomConstructed);
  document.head.appendChild(df);
  assert_true(customConstructed);
  assert_true(isCustomConstructed);
});
</script>
